removed user from getters and added own profile

+++ backend/src/controller/client.js
@@ -1,4 +1,4 @@
-const { save, getAll, getOne, deleteByUserId } = require('../service/client');
+const { save, getAll, getOne, getOwnByUserId, deleteByUserId } = require('../service/client');
 
 exports.save = async (req, res, next) => {
     try {
@@ -30,6 +30,16 @@ exports.getOne = async (req, res, next) => {
 	} catch (err) {
 		next(err);
 	}
+}
+
+exports.getOwn = async (req, res, next) => {
+    try {
+        const { status, payload } = await getOwnByUserId(req.user.id);
+
+        res.status(status).json(payload);
+    } catch (err) {
+        next(err);
+    }
 }
 
 exports.deleteOne = async (req, res, next) => {
